add all scc 3.7.0 flags; weighted_complexity is now double

+++ R/empty_scc_tibble.R
@@ -20,7 +20,7 @@ empty_scc_tibble <- function(by_file = FALSE) {
       comments            = integer(),
       blanks              = integer(),
       complexity          = integer(),
-      weighted_complexity = integer(),
+      weighted_complexity = double(),
       bytes               = integer(),
       uloc                = integer()
     )
@@ -34,7 +34,7 @@ empty_scc_tibble <- function(by_file = FALSE) {
       comments            = integer(),
       blanks              = integer(),
       complexity          = integer(),
-      weighted_complexity = integer(),
+      weighted_complexity = double(),
       bytes               = integer(),
       generated           = logical(),
       minified            = logical()
